+++ test/test_http_server.rb
@@ -34,7 +34,7 @@ class HTTP1ServerTest < MiniTest::Test
 
     response = connection.readpartial(8192)
     expected = <<~HTTP.chomp.crlf_lines.chomp
-      HTTP/1.1 200
+      HTTP/1.1 200 OK
       Content-Length: 13
 
       Hello, world!
@@ -52,7 +52,7 @@ class HTTP1ServerTest < MiniTest::Test
 
     response = connection.readpartial(8192)
     expected = <<~HTTP.crlf_lines.chomp
-      HTTP/1.1 200
+      HTTP/1.1 200 OK
       Content-Length: 13
 
       Hello, world!
@@ -67,14 +67,16 @@ class HTTP1ServerTest < MiniTest::Test
 
     connection << "GET / HTTP/1.0\r\nConnection: keep-alive\r\n\r\n"
     response = connection.readpartial(8192)
+    snooze
     assert !connection.eof?
-    assert_equal("HTTP/1.1 200\r\nContent-Length: 2\r\n\r\nHi", response)
+    assert_equal("HTTP/1.1 200 OK\r\nContent-Length: 2\r\n\r\nHi", response)
 
     connection << "GET / HTTP/1.1\r\n\r\n"
     response = connection.readpartial(8192)
+    snooze
     assert !connection.eof?
     expected = <<~HTTP.crlf_lines.chomp
-      HTTP/1.1 200
+      HTTP/1.1 200 OK
       Content-Length: 2
 
       Hi
@@ -83,8 +85,9 @@ class HTTP1ServerTest < MiniTest::Test
 
     connection << "GET / HTTP/1.0\r\n\r\n"
     response = connection.readpartial(8192)
+    snooze
     assert connection.eof?
-    assert_equal("HTTP/1.1 200\r\nContent-Length: 2\r\n\r\nHi", response)
+    assert_equal("HTTP/1.1 200 OK\r\nContent-Length: 2\r\n\r\nHi", response)
   end
 
   def test_pipelining_client
@@ -101,10 +104,10 @@ class HTTP1ServerTest < MiniTest::Test
     response = connection.readpartial(8192)
 
     expected = <<~HTTP.crlf_lines.chomp
-      HTTP/1.1 200
+      HTTP/1.1 200 OK
       Content-Length: 13
 
-      Hello, world!HTTP/1.1 200
+      Hello, world!HTTP/1.1 200 OK
       Content-Length: 14
 
       Hello, foobar!
@@ -191,9 +194,10 @@ class HTTP1ServerTest < MiniTest::Test
 
     connection << "GET / HTTP/1.1\r\n\r\n"
     response = connection.readpartial(8192)
+    snooze
     assert !connection.eof?
     expected = <<~HTTP.crlf_lines.chomp
-      HTTP/1.1 200
+      HTTP/1.1 200 OK
       Content-Length: 2
 
       Hi
@@ -209,6 +213,7 @@ class HTTP1ServerTest < MiniTest::Test
 
     snooze
     response = connection.readpartial(8192)
+    snooze
     assert !connection.eof?
     expected = <<~HTTP.crlf_lines
       HTTP/1.1 101 Switching Protocols